requirements.txt
jax==0.4.30
jaxlib==0.4.28
absl-py==2.1.0
matplotlib==3.9.0
ml_collections==0.1.1
wandb==0.16.0
scipy==1.12.0
flax==0.8.5
tabulate==0.9.0
optax==0.2.3
torch==2.3.0
git+https://github.com/haydn-jones/SOAP_JAX
psgd-jax==0.2.9